hackathon hack a penuer/New folder/requirements.txt
absl-py                      
asttokens                    
astunparse                   
async-generator              
attrs                        
backcall                     
beautifulsoup               
cachetools                   
certifi                      
cffi                         
chardet                      
charset-normalizer           
colorama                     
comm                         
contourpy                    
cycler                       
debugpy                      
decorator                    
exceptiongroup               
executing                    
filelock                     
flatbuffers                  
fonttools                    
gast                         
google-api-core              
google-api-python-client     
google-auth                  
google-auth-httplib         
google-auth-oauthlib         
google-cloud-core            
google-cloud-speech          
google-cloud-translate       
google-pasta                 
googleapis-common-protos     
googletrans                  rc
grpcio                       
grpcio-status                
h                          
h                           
hpy                         
hpack                        
hstspreload                  
httpcore                     
httplib                     
httpx                        
hyperframe                   
idna                         
imageio                      
imageio-ffmpeg               
ipykernel                    
ipython                      
jax                          
jedi                         
Jinja                       
joblib                       
jupyter_client               
jupyter_core                 
keras                        
kiwisolver                   
libclang                     
lxml                         
Markdown                     
MarkupSafe                   
matplotlib                   
matplotlib-inline            
ml-dtypes                    
moviepy                      
mpmath                       
nest-asyncio                 
networkx                     
numpy                        
oauthlib                     
opencv-python                
opt-einsum                   
outcome                      
packaging                    
pandas                       
parso                        
pickleshare                  
Pillow                       
pip                          
platformdirs                 
playsound                    
proglog                      
prompt-toolkit               
proto-plus                   
protobuf                     
psutil                       
pure-eval                    
pyasn                       
pyasn-modules               
pycparser                    
pydub                        
Pygments                     
pyparsing                    
PySocks                      
python-dateutil              
pytube                       
pytz                         
pywin                      
PyYAML                       
pyzmq                        
requests                     
requests-oauthlib            
rfc                      
rsa                          
scikit-learn                 
scipy                        
seaborn                      
selenium                     
setuptools                   
six                          
sniffio                      
sortedcontainers             
soupsieve                    
SpeechRecognition            
stack-data                   
sympy                        
tensorboard                  
tensorboard-data-server      
tensorflow                   
tensorflow-estimator         
tensorflow-intel             
tensorflow-io-gcs-filesystem 
termcolor                    
threadpoolctl                
torch                        
torchvision                  
tornado                      
tqdm                         
traitlets                    
trio                         
trio-websocket               
typing_extensions            
tzdata                       
uritemplate                  
urllib                      
wcwidth                      
Werkzeug                     
wheel                        
wrapt                        
wsproto                      
youtube-dl                   